ignore case of response header names so app override always succeeds
Header names are case insensitive as per e.g. RFC 7230
and `Response$header()` records them as-is.
This is fine, *except for* `Content-Type` and `Content-Length`, which
are treated specially by `Rhttpd`, and where this treatment fails if
app's choice of capitalization doesn't match `Rhttpd`'s:

e.g. a Rook app using this fragment:
```R
    ##...
    res$header("Content-type", "application/json")
    ##...
```
still leads to the client getting a response of Content-Type "text/html",
because lookup in `Rhttpd` is by the differently-capitalized `Content-Type`